LibWeb: Use handle_failed_fetch to implement handle_failed_decode
...When loading images through SharedImageRequest.

There is a small behavioural difference here - handle_failed_fetch
clears the pending callbacks whereas handled_failed_decode was
previosuly not. This does not seem intentional, and appears like a bug.

Implementing it this was is simpler - and also means we don't need to
take a strong handle to this in the case of loading an SVG image.