LibWeb/HTML: Use helper method for "enabling port's port message queue"
After a long reading of the spec, I still can't quite figure out what
the normative impact of enabling this message queue, so let's keep it
as is until someone is able to fully understand it. But it does allow
us to follow the wording of the spec a little more closely, as this
concept is used on MessagePorts outside of the implementation of
MessagePort's start() method.