LibWeb: Don't crash on document.domain setter for domain not in PSL
The specification algorithm to get a public suffix from a host
ensures that a host with a domain always returns a public
suffix. URL::get_public_suffix was not doing that, while
URL::Host::public_suffix does (albiet, incorrectly). This fixes
a crash seen on WPT when the document.domain setter is called on
a document which is hosted from a domain which is not present in
the public suffix list.